Add trace cofiguration manager
This commit adds the TraceConfigurationManager to handle calls to the
trace server to manage configurations in the configuration UI.
